feat: Implement dashboard screen with biometric charts
Introduces the main dashboard screen, which fetches and displays biometric data (heart rate and respiratory rate) from a repository. This feature includes setting up the necessary data flow from the repository to the ViewModel and finally to the Compose UI.

Key changes in this commit:

- **Dashboard UI (`DashboardScreen.kt`)**:
    - Creates a Jetpack Compose screen to display average heart rate and respiratory rate.
    - Integrates the Vico charting library to visualize the heart rate series as a column chart.
    - Implements conditional rendering to show a loading indicator, an error message, or the data and chart based on the current UI state. This robustly handles the initial empty state and prevents crashes when the chart data is not yet available.

- **ViewModel (`DashboardViewModel.kt`)**:
    - Implements the `refresh()` function to fetch data from the `BiometricsRepository` concurrently using Kotlin coroutines (`async`/`await`).
    - Manages UI state using a `StateFlow`, providing updates for loading, success, and error states.
    - Improves error handling and data flow by using `runCatching` to ensure the UI state is updated atomically and safely.

- **Repository (`MockBiometricsRepository.kt`)**:
    - Fixes type mismatches by correctly converting `Double` values from the API to the `Float` type expected by the Room database entities (`SampleEntity` and `SummaryEntity`).
    - Correctly uses Float literals (`0.0f`) for default values to match the entity definitions.

+++ app/src/main/java/com/yuli/svastha/data/db/Entities.kt
@@ -7,12 +7,12 @@ data class SampleEntity(
   @PrimaryKey(autoGenerate = true) val id: Long = 0,
   val metric: String, // "heartRateBpm" | "respiratoryRateBpm"
   val ts: String,
-  val value: Double
+  val value: Float
 )
 
 @Entity(tableName = "daily_summary")
 data class SummaryEntity(
   @PrimaryKey val date: String,
-  val hrMin: Double, val hrAvg: Double, val hrMax: Double, val hrRest: Double,
-  val rrMin: Double, val rrAvg: Double, val rrMax: Double
+  val hrMin: Float, val hrAvg: Float, val hrMax: Float, val hrRest: Float,
+  val rrMin: Float, val rrAvg: Float, val rrMax: Float
 )
